+++ catalog/files/impl/src/main/java/org/projectnessie/catalog/files/s3/S3ClientSupplier.java
@@ -38,6 +38,7 @@ import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.DelegatingS3Client;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.S3ClientBuilder;
+import software.amazon.awssdk.services.s3.S3Configuration;
 import software.amazon.awssdk.services.s3.model.S3Request;
 import software.amazon.awssdk.services.s3.model.S3Request.Builder;
 
@@ -89,7 +90,7 @@ public class S3ClientSupplier {
             .overrideConfiguration(
                 override -> override.defaultProfileFileSupplier(() -> EMPTY_PROFILE_FILE))
             .serviceConfiguration(
-                serviceConfig -> serviceConfig.profileFile(() -> EMPTY_PROFILE_FILE));
+                serviceConfig -> configureServiceConfiguration(bucketOptions, serviceConfig));
 
     if (LOGGER.isTraceEnabled()) {
       LOGGER.trace(
@@ -114,6 +115,12 @@ public class S3ClientSupplier {
     }
 
     return s3Client;
+  }
+
+  static void configureServiceConfiguration(
+      S3BucketOptions bucketOptions, S3Configuration.Builder serviceConfig) {
+    serviceConfig.profileFile(() -> EMPTY_PROFILE_FILE);
+    bucketOptions.chunkedEncodingEnabled().ifPresent(serviceConfig::chunkedEncodingEnabled);
   }
 
   private static String toLogString(S3BucketOptions options) {